Fix comments in MR
Changes in this commig:
- Files move upper
- Add comments to exported functions
- Span is propagated through context
- Code cleanup
- Sort imports
