+++ contrib/go-pg/pg_go.go
@@ -1,0 +1,52 @@
+package pg
+
+import (
+	"context"
+	"time"
+
+	"github.com/go-pg/pg/v10"
+	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace"
+	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
+	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
+)
+
+const gopgStartSpan = "dd-trace-go:span"
+
+// ADd QueryHook to existing go_pg.DB instance
+func Hook(db *pg.DB) *pg.DB {
+	db.AddQueryHook(&QueryHook{})
+	return db
+}
+
+// QueryHook for go_pg
+type QueryHook struct{}
+
+// This hook is executed before query is sent
+// Start measure, when query is started
+func (h *QueryHook) BeforeQuery(ctx context.Context, qe *pg.QueryEvent) (context.Context, error) {
+	if qe.Stash == nil {
+		qe.Stash = make(map[interface{}]interface{})
+	}
+
+	unformatedSql, _ := qe.UnformattedQuery()
+
+	opts := []ddtrace.StartSpanOption{
+		tracer.StartTime(time.Now()),
+		tracer.SpanType(ext.SpanTypeSQL),
+		tracer.ResourceName(string(unformatedSql)),
+	}
+	_, ctx = tracer.StartSpanFromContext(ctx, "gopg", opts...)
+	return ctx, qe.Err
+}
+
+// Hook is executed after query is finished
+func (h *QueryHook) AfterQuery(ctx context.Context, qe *pg.QueryEvent) error {
+	span, ok := tracer.SpanFromContext(ctx)
+	if !ok {
+		return nil
+	}
+
+	span.Finish(tracer.WithError(qe.Err))
+
+	return qe.Err
+}
