+++ contrib/go-pg/pg_go_test.go
@@ -1,18 +1,17 @@
-package pggo
+package pg
 
 import (
 	"context"
-	"fmt"
+	"testing"
+
 	"github.com/go-pg/pg/v10"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/mocktracer"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
-	"testing"
 )
 
-
-func GetPostgresConnection() *pg.DB{
+func GetPostgresConnection() *pg.DB {
 	db := pg.Connect(&pg.Options{
 		User:     "postgres",
 		Password: "",
@@ -21,7 +20,7 @@ func GetPostgresConnection() *pg.DB{
 	return db
 }
 
-func TestSelect(t *testing.T){
+func TestSelect(t *testing.T) {
 	assert := assert.New(t)
 	mt := mocktracer.Start()
 	defer mt.Stop()
@@ -29,7 +28,7 @@ func TestSelect(t *testing.T){
 	conn := GetPostgresConnection()
 	Hook(conn)
 
-	parentSpan, ctx := tracer.StartSpanFromContext(context.TODO(), "http.request",
+	parentSpan, ctx := tracer.StartSpanFromContext(context.Background(), "http.request",
 		tracer.ServiceName("fake-http-server"),
 		tracer.SpanType(ext.SpanTypeWeb),
 	)
@@ -39,8 +38,7 @@ func TestSelect(t *testing.T){
 	parentSpan.Finish()
 	spans := mt.FinishedSpans()
 
-	assert.True(len(spans) >= 2)
-	fmt.Println(spans[0].OperationName())
+	assert.True(len(spans) == 2)
 	assert.Equal("gopg", spans[0].OperationName())
 	assert.Equal("http.request", spans[1].OperationName())
 }
